Pull changes from Juniper/web-controller to absingla/web-controller

+++ webroot/config/physicaldevices/physicalinterfaces/ui/js/physicalinterfaces_config.js
@@ -355,18 +355,42 @@ function physicalInterfacesConfig() {
     function deletePhysicalInterface(selected_rows) {
         $('#btnDeletePhysicalInterface').addClass('disabled-link');	
         if(selected_rows && selected_rows.length > 0){
-            var deleteAjaxs = [];
+            var physicalIntfsDeleteAjaxs = [];
+            var logicalIntfsDeleteAjaxs = [];
             for(var i = 0;i < selected_rows.length;i++){
                 var sel_row_data = selected_rows[i];
-                deleteAjaxs[i] = $.ajax({
-                    url:'/api/tenants/config/physical-interface/' + currentUUID + '/' + sel_row_data['type'] + '/' + sel_row_data['uuid'],
-                    type:'DELETE'
-                });
+                if(sel_row_data['type'] == 'Logical'){
+                    logicalIntfsDeleteAjaxs.push($.ajax({
+                        url:'/api/tenants/config/physical-interface/' + currentUUID + '/' + sel_row_data['type'] + '/' + sel_row_data['uuid'],
+                        type:'DELETE'
+                    }));
+                } 
             }
-            $.when.apply($,deleteAjaxs).then(
+            //First delete logical interfaces
+            $.when.apply($,logicalIntfsDeleteAjaxs).then(
                 function(response){
-                    //all success
-                    fetchData();
+                    //now delete physical interfaces
+                    for(var i = 0;i < selected_rows.length;i++){
+                        var sel_row_data = selected_rows[i];
+                        if(sel_row_data['type'] == 'Physical'){
+                            physicalIntfsDeleteAjaxs.push($.ajax({
+                                    url:'/api/tenants/config/physical-interface/' + currentUUID + '/' + sel_row_data['type'] + '/' + sel_row_data['uuid'],
+                                    type:'DELETE'
+                                }));
+                        } 
+                    }
+                    $.when.apply($,physicalIntfsDeleteAjaxs).then(
+                            function(response){
+                                //all success
+                                fetchData();
+                            },
+                            function(){
+                                //if at least one delete operation fails
+                                var r = arguments;
+                                showInfoWindow(r[0].responseText,r[2]);     
+                                fetchData();
+                            }
+                        );
                 },
                 function(){
                     //if at least one delete operation fails
@@ -375,7 +399,26 @@ function physicalInterfacesConfig() {
                     fetchData();
                 }
             );
+            if(deleteVMIs.length > 0) {
+                deleteVirtulMachineInterfaces(deleteVMIs);
+            }
         }
+    }
+
+    function deleteVirtulMachineInterfaces(deleteVMIs) {
+        var deleteAjaxs = [];
+        for(var i = 0; i < deleteVMIs.length; i++) {
+            deleteAjaxs[i] =  $.ajax({
+                    url:'/api/tenants/config/delete-port/' + deleteVMIs[i],
+                    type:'DELETE'
+                })
+        }
+        $.when.apply($,deleteAjaxs).then(
+            function(response){
+            },
+            function(){
+            }
+        );
     }    
     
     window.physicalInterfaceEditWindow = function(index) {
@@ -519,9 +562,12 @@ function physicalInterfacesConfig() {
                 }
             }
         }
-        doAjaxCall(url, methodType, JSON.stringify(postObject), 'successHandlerForPhysicalInterfaces', 'failureHandlerForPhysicalInterfaces', null, null);
+        doAjaxCall(url, methodType, JSON.stringify(postObject), 'successHandlerForPhysicalInterfaces', 'failureHandlerForCreatePhysicalInterfaces', null, null);
     }
     
+    function failureHandlerForCreatePhysicalInterfaces(error) {
+        fetchData();
+    }
     function clearCreateEditWindow() {
         $('#ddType').data('contrailDropdown').value('logical');
         $('#txtPhysicalInterfaceName').removeAttr('disabled');
@@ -730,7 +776,8 @@ function physicalInterfacesConfig() {
                     server : liDetails.vmiDetails != null ? liDetails.vmiDetails : '-',
                     vn : liDetails.vnRefs != null ? liDetails.vnRefs : '-',
                     li_type : liDetails.liType != null ? liDetails.liType : '-' ,
-                    vmi_ip : liDetails.vmiIP != null ? liDetails.vmiIP : '-'
+                    vmi_ip : liDetails.vmiIP != null ? liDetails.vmiIP : '-',
+                    vmi_uuid : liDetails.vmiUUID
                 });
                 var lInterfaces = pInterfaces[i]['physical-interface'] ? pInterfaces[i]['physical-interface']['logical_interfaces'] : null;
                 var lInterfaceNames = '';
@@ -754,7 +801,8 @@ function physicalInterfacesConfig() {
                             server : liDetails.vmiDetails,
                             vn : liDetails.vnRefs,
                             li_type : liDetails.liType,
-                            vmi_ip : liDetails.vmiIP
+                            vmi_ip : liDetails.vmiIP,
+                            vmi_uuid : liDetails.vmiUUID
                         });                        
                     }
                     var currPhysicalInfRow = getCurrentPhysicalInfRow(gridDS, pInterface.uuid);
@@ -820,19 +868,20 @@ function physicalInterfacesConfig() {
             liType = liType === 'l2' ? 'L2' : 'L3';
         }
         var vmiDetails = inf['vmi_details'] ? inf['vmi_details'].mac[0] +' ('+ inf['vmi_details'].ip[0] + ')' : '-';
+        var vmiUUID;
         if(vmiDetails != '-') {
             vmiIP = inf['vmi_details'].ip[0];
             vnRefs = inf['vmi_details']['vn_refs'] ? inf['vmi_details']['vn_refs'][0].to : '-';
             if(vnRefs != '-') {
                 vnRefs = vnRefs[2] + ' (' + vnRefs[0] + ':' + vnRefs[1] + ')';
-            }                            
+            }
+            vmiUUID = inf['vmi_details']['vmi_fq_name'][2];
         }
-        return { vlanTag : vlanTag, liType : liType, vmiDetails : vmiDetails, vnRefs : vnRefs, vmiIP : vmiIP};
+        return { vlanTag : vlanTag, liType : liType, vmiDetails : vmiDetails, vnRefs : vnRefs, vmiIP : vmiIP, vmiUUID : vmiUUID};
     }
     
     window.failureHandlerForPhysicalInterfaces =  function(error) {
-         //gridPhysicalInterfaces.showGridMessage("errorGettingData");
-         fetchData();
+         gridPhysicalInterfaces.showGridMessage("errorGettingData");
     }
     
     function validate() {
